+++ app/src/main/java/com/statfinder/statfinder/QuestionActivityFromSearch.java
@@ -76,15 +76,15 @@ public class QuestionActivityFromSearch extends FragmentActivity {
                         totalRef.addListenerForSingleValueEvent(new ValueEventListener() {
                             @Override
                             public void onDataChange(DataSnapshot votesSnapshot) {
-                                float totalFlags = (float) flagSnapshot.getValue();
-                                float totalVotes = (float) votesSnapshot.getValue();
-                                float totalInteractions = totalFlags + totalVotes;
+                                double totalFlags =  ((Long) flagSnapshot.getValue()).doubleValue();
+                                double totalVotes =  ((Long) votesSnapshot.getValue()).doubleValue();
+                                double totalInteractions = totalFlags + totalVotes;
                                 if (totalInteractions > 10 && totalInteractions < 20) {
                                     if (totalFlags > totalVotes) {
                                         ref.removeValue();
                                     }
                                 } else if (totalInteractions > 20) {
-                                    float percentRage = totalFlags / totalInteractions;
+                                    double percentRage = totalFlags / totalInteractions;
                                     if (percentRage > 0.25) {
                                         ref.removeValue();
                                         Firebase userRef = new Firebase("https://statfinderproject.firebaseio.com/Users/" + creator + "/CreatedQuestions/" + questionID);
